backend/edu/requirements.txt
Django==2.2.8
djangorestframework==3.11.0
Markdown==3.1.1
django-filter==2.2.0
django-allauth==0.41.0
graphene-django==2.8.0
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1
celery==4.4.0
redis==3.3.11
pytest==5.3.2
pytest-django==3.7.0
coverage==5.0.2
drf-yasg==1.17.0
sentry-sdk==0.13.5
ipython==7.11.1
